requirements.txt
In this file, there is a list of packages that can be needed to execute the codes for the thesis.
Including versions of the packages.

backtrader==1.9.76.123
Cython==0.29.28
distfit==1.5.5
fitter==1.4.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.2
pandas==1.3.3
pmdarima==1.8.5
python-dateutil==2.8.2
scipy==1.9.3
seaborn==0.11.2
seaborn-qqplot==0.5.0
statsmodels==0.13.2

Packages related to notebooks:
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.3
jupyter-core==4.8.1
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
notebook==6.4.4
notebook-shim==0.1.0